BDIM clean up
I did a large clean up of the BDIM file. I figured if it was too long and confusing for the author to understand, then it could stand some revision.

I added a better write up and example code to the top

I consolidated the u and p update into a new routine updateUP and gave that routine a little write up as well

This leaves the update and update2 routines nice and clear and help differentiates between QUICK and SEMILAGRANGIAN

I also took out the get_coeff call in update2 since it is not needed