Generalized `react` functions and area calculations
I created a react function in the parent class `Body` and removed
the overlapping `react` functions in CicleBody and NACA.

I added function getArea to class `Body` to compute the area for
any general convex polygon body. If the body is closed then this
function is called and the area is saved.

I replaced the area formula in CircleBody with a more general formula
in EllipseBody.

I corrected a typo in the `FoilArea` parameter and then commented it out
since the getArea function was within 1% of this value. I adjusted the `mr`
parameter in the example code to get the same final mass.

I added parameter `mass` to Body and replaced `mr*area` with `mass`
in the F=ma intgral. This allows you to set the mass for bodies
without a well defined area (like open bodies). The default for a
body with non-zero area is `mass=area`.

I deleted the parameters `mr` and `dxc2` from `Body` as they
were more appropriate as local variables, not saved parameters.

Added parameters `x/y free` to `Body` which can be set to disable
freedom of motion in either direction. The default is free, since
calling the react() function, should default to moving the body.