Free rotational capability
Added moment of inertia and coordinate-based calculation in body.
Added pressMoment routine to find the integrated pressure moment.
Added rotation to react function.

!! Removed the ability to pass the mass ratio to react. Instead, please
set the mass and/or inertia in setup(). I have adjusted the FreeCylinder
and AmandaTest classes to follow this recommendation. !!

Turned off the ellipse-specific approximations for the distance
and normal functions to allow rotation.