"Fix" for adaptive time step
Moved the update of the pressure coefficients into BDIM.update. This ensures
that these coefficients are recomputed and the pressure force is correct.

It is very unlikely that this was a problem for anyone. The only way the
coefficients were not computed is if a body was not passed to BDIM.update
AND adaptive time step was active. Even then - the flow is still
correct, the forces are just scaled by dt/dt_initial.