Cleaned up Body
A set of case specific functions had been added to Body. I made a new class
AmandaNACA and mode the changes into that class.

This is good coding practice. When you are adding functionality ask yourself if
the changes are general or problem specific. If they are likely to be useful for
projects other than yours, then add them as deep as possible to the class. If they
are very specific to your problem or your set up (like how you process data) then
add a new class with these features.

But don't stress - we can always move stuff later.