Body.update behavoir changed to be more intuitive
Body.update now zeros out velocity data if no changes to the position
and angle have been made since the last time Body.update was called.
This more strongly links the position and velocity boundary conditions
and matches the intuitive expectation that the velocity BC should be
zero if the body was not moved.

To compliment this new behavoir, a new method Body.updatePositionOnly
was added. If this is used instead of body.update, the position changes
will not enforce a non-zero BC. This is useful for intializing body
positions or performing pseudo-steady runs.