Improved MG tolerance
Implemented the improved tolerance check from Lotus.

Bascially, the check before was not a good indicator of
solver average error. It over-penalized large grids.
The new version ensures the average residual is less
than 0.01%, which is plenty.

This may significantly speed up your code! Especially on
large grids!! So make sure to pull/merge this change.