Add motion path
This update is in responce to a bug found by Dr Mingyi Tan at Southampton.
The bug was due to the user easily overcomming the CFL limit when
interactively moving the default sketch. If the user defined translation
was extremely large the code often damped it out (unintentionally) but
if it was around 2 times the CFL limit, the flow would blow up.

Instead of adding (intentional) damping, I added a layer between the user
and the body motion. A path ArrayList is used to store the motions
prescribed by the user, which are then applied to the body as fast as the
CFL limit allows. The path is also displayed with the body.

The result is visually pleasing and more accurate (no more numerical
damping or explosions).