Update body motion routines
The body.update() function was confusing and caused problems.
I have replaced it with body.follow(). If you pass a PVector
to follow() then the body will move to that position and
automatically compute the correct velocity. If you don't pass
an argument, it will follow the mouse when pressed.

The function initPath() lets you set the position without
inducing a velocity.

I've included two nice clear demos of this in the main
LilyPad.pde file. I think 95% of projects should be using
follow() instead of translate and rotate directly.

The body.unsteady flag has been changed to a function which
avoids the need for body.update() for most classes. Instead
body.unstady() can be hacked as needed.

I've tried to update all the files, but I may have missed
a few. I've also removed a few files that were attempts at
similar solutions.