Add btype=3 for z-vectors
Included a new field type for z-vectors, which occur when taking
the curl of a VectorField, such as the vorticity.

Changed VectorField.vorticity() to VectorField.curl() to be more
general and return the new field type. The values of these fields
are located at the corners of the cells. This reduces the amount
of interpolation needed.

Added Field.curl() to take the curl of z-vectors (only) and return
a VectorField. The routines now obey an important identity
u.curl().curl() = -u.laplacian() if u.divergence = 0.

Adjusted the interpolation functions in Field and used the new
function name throughout the pde files.