Update BodyUnion velocity/normal
The velocity and normal vector from BodyUnion were discontinuous for
intersecting bodies - which leads to discontinuous pressure and
forces, especially when the body moves through grid points.

I added a weight function which smoothly combined the velocity and
normal vectors, leading to a continuous set of fields.

Caution:
It is possible that even if both of the component velocity fields
are divergence free, this smoothed field will not be.