Add derivative kinematics to follow
Added the ability to pass the dx=u*dt conditions directly to follow.
This eliminates the need for initPath, just use follow(x,0). It also
eliminates a bug for very high resolution and large amplitude motion.
In that circumstance, the automatic backwards difference version of
the velocity was not smooth enough, leading to large amplitude noise.