Update body.react and new ChaoticEllipse example
This update generalizes the react function to robustly handle
general 3DOF motion. This function integrates the force on the
body to update the velocity (O(1)) and position (O(2))

The function uses the added mass coefficients to keep the integration
stable, and these should be supplied for any new body geometries!
Those coefficents have been added for an ellipse (and circle) and the
ellipse values have been added as an estimate for the NACA foil.

To demonstrate this capability, the FreeCylinder class has been
replaced by ChaoticEllipse, based on the Weymouth 2014 paper. There
are also exmaple modes demonstrating a mix of free and actuated
motion, and imposing constraints.