Non-convex body update
Added the capability to determine the distance function to
general non-convex bodies. The method combines the |distance|
function to the segments defining the shape and then uses
the winding number to determine the sign of the distance.

Removed the convex enforcement in body.add() and included a
convexity check in body.end(). Based on this, the correct
distance function is chosen automatically.