Added WaveyDuncan
BC flag waveInlet was added to Field which is the same as gradientExit
but also avoids overwritting the inlet velocity.

Class WaveyDuncan extends Duncan by adding a linear wave velocity
profile to the inlet bval. The function takes in the dimensionless
wave number kL and wave steepness ak, but I haven't checked how
well these correspond to the measured values.