Major update to ImageBody and new CamBody class
The heart of ImageBody is really the brightness array, which is now stored directly in `pix`. This avoids the need to maintain the correct colorMode while running a simulation.

CamBody extends an ImageBody from a camera capture source. Use `createCamBody`. New methods `update` grabs the newest capture and `background` uses that capture as the simulation background.