Make Body and PixelBody both extend AbstractBody
Added a new AbstractBody class to be the parent of both Body and PixelBody.

This let me get rid of the custom code in BDIM to handle different body classes.

Required me to move the zeroth and first moment stuff in Body.pde, but it's cleaner now.

PixelBody has its own method for the moments, but the first moment, delta1, is still a bit of a hack.

Finally, a CamBody should update every step, so I've set unsteady=true. Right now AbstractBody defaults to unsteady=false, and PixelBody doesn't mod it.